+++ lib/simple_markdown_editor.dart
@@ -1,5 +1,11 @@
 library simple_markdown_editor;
 
+export 'widgets/markdown_form_field.dart';
+export 'widgets/markdown_parse.dart';
+export 'widgets/markdown_toolbar.dart';
+export 'widgets/markdown_field.dart';
+
+// deprecated
 export 'widgets/z_markdown_editor.dart';
 export 'widgets/z_markdown_parse.dart';
 export 'widgets/z_markdown_toolbar.dart';